Extract emulator states into a separate module
Google Cloud saves don't support save RAM persistence, but save state (.dat) files only.
