+++ pkg/emulator/libretro/nanoarch/naemulator.go
@@ -12,7 +12,7 @@ import (
 
 	config "github.com/giongto35/cloud-game/v2/pkg/config/emulator"
 	"github.com/giongto35/cloud-game/v2/pkg/emulator"
-	"github.com/giongto35/cloud-game/v2/pkg/util"
+	"github.com/giongto35/cloud-game/v2/pkg/persistence"
 )
 
 /*
@@ -268,9 +268,9 @@ func (na *naEmulator) ToggleMultitap() error {
 	return nil
 }
 
-func (na *naEmulator) GetHashPath() string { return util.GetSavePath(na.roomID) }
+func (na *naEmulator) GetHashPath() string { return persistence.GetMainState(na.roomID) }
 
-func (na *naEmulator) GetSRAMPath() string { return util.GetSRAMSavePath(na.roomID) }
+func (na *naEmulator) GetSRAMPath() string { return persistence.GetSavePath() + na.roomID + ".srm" }
 
 func (*naEmulator) GetViewport() interface{} {
 	return outputImg
