+++ pkg/emulator/libretro/nanoarch/persistence_test.go
